# apps/api/openapi.json
{
 "openapi": "3.0.3",
 "info": {
  "title": "Intuitive Fusion Procurement Portal API (POC)",
  "version": "0.1.0-draft",
  "description": "REST API for the POC. Auth is mock (HttpOnly session cookie) behind an IdentityProvider interface. All mutations emit audit events. Errors follow RFC 7807 (application/problem+json) with stable `code` and `correlationId`."
 },
 "servers": [
  {
   "url": "http://localhost:4000/api/v1",
   "description": "Local dev"
  }
 ],
...
 "security": [
  {
   "sessionCookie": []
  }
 ],
 "paths": {
...
  "/suppliers": {
   "get": {
    "operationId": "listSuppliers",
    "tags": [
     "SupplierPortal"
    ],
    "summary": "Supplier directory with sanctions and insurance status",
    "parameters": [],
    "x-roles": [
     "PROCUREMENT",
     "LEGAL",
     "FINANCE",
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/Supplier"
         }
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   }
  },
  "/suppliers/{id}": {
   "get": {
    "operationId": "getSupplier",
    "tags": [
     "SupplierPortal"
    ],
    "summary": "Supplier profile: status, contacts, tenders bid on, contracts held",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "PROCUREMENT",
     "LEGAL",
     "FINANCE",
     "ADMIN"
    ],
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SupplierProfile"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     }
    }
   }
  },
  "/suppliers/{id}/contacts": {
   "post": {
    "operationId": "addSupplierContact",
    "tags": [
     "SupplierPortal"
    ],
    "summary": "A buyer adds a contact to an existing supplier; returns a one-time activation link (no email is sent in the proof of concept)",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": [
     "PROCUREMENT"
    ],
    "responses": {
     "201": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SupplierContactAdded"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "401": {
      "$ref": "#/components/responses/E401"
     },
     "403": {
      "$ref": "#/components/responses/E403"
     },
     "404": {
      "$ref": "#/components/responses/E404"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "description": "409 EMAIL_IN_USE",
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SupplierContactCreate"
       }
      }
     }
    }
   }
  },
  "/supplier/activate/{token}": {
   "get": {
    "operationId": "getActivation",
    "tags": [
     "SupplierPortal"
    ],
    "summary": "Look up an activation link (one generic 404 for unknown, used or expired)",
    "parameters": [
     {
      "name": "token",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ],
    "x-roles": "public",
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ActivationInfo"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "429": {
      "$ref": "#/components/responses/E429"
     }
    },
    "security": []
   }
  },
  "/supplier/activate": {
   "post": {
    "operationId": "activateContact",
    "tags": [
     "SupplierPortal"
    ],
    "summary": "Set a password with a one-time activation link",
    "parameters": [],
    "x-roles": "public",
    "responses": {
     "200": {
      "description": "Success",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Message"
        }
       }
      }
     },
     "400": {
      "$ref": "#/components/responses/E400"
     },
     "429": {
      "$ref": "#/components/responses/E429"
     },
     "409": {
      "$ref": "#/components/responses/E409"
     },
     "422": {
      "$ref": "#/components/responses/E422"
     }
    },
    "security": [],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ActivationRequest"
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "sessionCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "if_session"
   }
  },
  "schemas": {
   "Problem": {
    "type": "object",
    "properties": {
     "type": {
      "type": "string"
     },
     "title": {
      "type": "string"
     },
     "status": {
      "type": "integer"
     },
     "detail": {
      "type": "string"
     },
     "code": {
      "type": "string"
     },
     "correlationId": {
      "type": "string"
     },
     "errors": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "field": {
         "type": "string"
        },
        "message": {
         "type": "string"
        }
       },
       "additionalProperties": false,
       "required": [
        "field",
        "message"
       ]
      }
     }
    },
    "additionalProperties": false,
    "required": [
     "type",
     "title",
     "status",
     "code",
     "correlationId"
    ]
   },
...
   "Message": {
    "type": "object",
    "properties": {
     "message": {
      "type": "string"
     }
    },
    "additionalProperties": false,
    "required": [
     "message"
    ]
   },
...
   "Supplier": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "company": {
      "type": "string"
     },
     "abn": {
      "type": "string"
     },
     "sanctionsStatus": {
      "type": "string",
      "enum": [
       "PENDING",
       "CLEAR",
       "MATCH"
      ]
     },
     "insuranceStatus": {
      "type": "string",
      "enum": [
       "UNKNOWN",
       "CURRENT",
       "EXPIRING",
       "EXPIRED"
      ]
     },
     "lastCheckedAt": {
      "type": "string",
      "format": "date-time"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "company",
     "sanctionsStatus"
    ]
   },
...
   "SupplierContact": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "name": {
      "type": "string"
     },
     "email": {
      "type": "string"
     },
     "active": {
      "type": "boolean"
     },
     "awaitingActivation": {
      "type": "boolean"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "name",
     "email"
    ]
   },
   "SupplierProfile": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "company": {
      "type": "string"
     },
     "abn": {
      "type": "string"
     },
     "sanctionsStatus": {
      "type": "string"
     },
     "insuranceStatus": {
      "type": "string"
     },
     "lastCheckedAt": {
      "type": "string",
      "format": "date-time"
     },
     "contacts": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/SupplierContact"
      }
     },
     "tenders": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "tenderId": {
         "type": "string",
         "format": "uuid"
        },
        "number": {
         "type": "string"
        },
        "title": {
         "type": "string"
        },
        "submission": {
         "type": "string"
        }
       },
       "additionalProperties": false
      }
     },
     "contracts": {
      "type": "array",
      "items": {
       "type": "object",
       "properties": {
        "id": {
         "type": "string",
         "format": "uuid"
        },
        "number": {
         "type": "string"
        },
        "status": {
         "type": "string"
        },
        "value": {
         "type": "number"
        }
       },
       "additionalProperties": false
      }
     },
     "canAddContact": {
      "type": "boolean"
     }
    },
    "additionalProperties": false,
    "required": [
     "id",
     "company"
    ]
   },
   "SupplierContactCreate": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "email": {
      "type": "string",
      "format": "email"
     }
    },
    "additionalProperties": false,
    "required": [
     "name",
     "email"
    ]
   },
   "SupplierContactAdded": {
    "type": "object",
    "properties": {
     "contact": {
      "$ref": "#/components/schemas/SupplierContact"
     },
     "activationPath": {
      "type": "string"
     },
     "expiresAt": {
      "type": "string",
      "format": "date-time"
     }
    },
    "additionalProperties": false,
    "required": [
     "contact",
     "activationPath"
    ]
   },
   "ActivationInfo": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "email": {
      "type": "string"
     },
     "company": {
      "type": "string"
     },
     "organisation": {
      "type": "string"
     },
     "expiresAt": {
      "type": "string",
      "format": "date-time"
     }
    },
    "additionalProperties": false
   },
   "ActivationRequest": {
    "type": "object",
    "properties": {
     "token": {
      "type": "string"
     },
     "password": {
      "type": "string",
      "minLength": 12
     }
    },
    "additionalProperties": false,
    "required": [
     "token",
     "password"
    ]
   },
...
  },
  "responses": {
   "E400": {
    "description": "Validation error",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E401": {
    "description": "Not authenticated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E403": {
    "description": "Forbidden",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E404": {
    "description": "Not found",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E409": {
    "description": "Conflict / rule violated",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
   "E422": {
    "description": "Business rule failed",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   },
...
   "E429": {
    "description": "Rate limited",
    "content": {
     "application/problem+json": {
      "schema": {
       "$ref": "#/components/schemas/Problem"
      }
     }
    }
   }
  }
 }
}